+++ exchange/src/shared/util.hpp
@@ -9,7 +9,7 @@ enum class SIDE { BUY, SELL };
 
 namespace util {
 constexpr bool
-is_close_to_zero(float value, float epsilon = 1e-6f)
+is_close_to_zero(double value, double epsilon = 1e-6f)
 {
     return std::fabs(value) < epsilon;
 }
